Revert " 🚀 Move FixedLayer to amp-viewer-integration extension (#27499)"
This reverts commit e3b0d953f16c7af4e261217872388d2c31599fa3.
