+++ .eslintrc
@@ -23,9 +23,9 @@
     "sourceType": "module"
   },
   "globals": {
-    "MOVE_FIXED_LAYER": "readonly",
     "NATIVE_CUSTOM_ELEMENTS_V1": "readonly",
     "IS_ESM": "readonly",
+
     "AMP": "readonly",
     "context": "readonly",
     "global": "readonly"
