+++ build-system/babel-config/test-config.js
@@ -16,7 +16,7 @@
 'use strict';
 
 const argv = require('minimist')(process.argv.slice(2));
-const {getReplacePlugin} = require('./helpers');
+const {getReplacePlugin} = require('./replace-plugin');
 
 /**
  * Gets the config for babel transforms run during `gulp [unit|integration]`.
