+++ build-system/global-configs/experiments-const.json
@@ -1,4 +1,3 @@
 {
-  "INTERSECTION_OBSERVER_POLYFILL_INABOX": false,
-  "MOVE_FIXED_LAYER": false
+  "INTERSECTION_OBSERVER_POLYFILL_INABOX": false
 }
