+++ extensions/amp-viewer-integration/0.1/amp-viewer-integration.js
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import {FixedLayer} from '../../../src/service/fixed-layer';
 import {FocusHandler} from './focus-handler';
 import {
   HighlightHandler,
@@ -88,11 +87,6 @@ export class AmpViewerIntegration {
 
     if (!this.isWebView_ && !origin) {
       return Promise.resolve();
-    }
-
-    if (MOVE_FIXED_LAYER) {
-      const viewport = Services.viewportForDoc(ampdoc);
-      viewport.createFixedLayer(FixedLayer);
     }
 
     if (this.isWebView_ || this.isHandShakePoll_) {
