+++ src/inabox/inabox-viewport.js
@@ -359,9 +359,6 @@ class InaboxViewportImpl {
   /** @override */
   removeFromFixedLayer(unusedElement) {}
 
-  /** @override */
-  createFixedLayer(unusedConstructor) {}
-
   /**
    * @private
    */
