+++ src/service/viewport/viewport-impl.js
@@ -139,12 +139,15 @@ export class ViewportImpl {
     /** @private {string|undefined} */
     this.originalViewportMetaString_ = undefined;
 
-    /** @private {?FixedLayer} */
-    this.fixedLayer_ = null;
-
-    if (!MOVE_FIXED_LAYER) {
-      this.createFixedLayer(FixedLayer);
-    }
+    /** @private @const {!FixedLayer} */
+    this.fixedLayer_ = new FixedLayer(
+      ampdoc,
+      this.vsync_,
+      this.binding_.getBorderTop(),
+      this.paddingTop_,
+      this.binding_.requiresFixedLayerTransfer()
+    );
+    ampdoc.whenReady().then(() => this.fixedLayer_.setup());
 
     this.viewer_.onMessage('viewport', this.updateOnViewportEvent_.bind(this));
     this.viewer_.onMessage('scroll', this.viewerSetScrollTop_.bind(this));
@@ -400,9 +403,6 @@ export class ViewportImpl {
 
   /** @override */
   isDeclaredFixed(element) {
-    if (!this.fixedLayer_) {
-      return false;
-    }
     return this.fixedLayer_.isDeclaredFixed(element);
   }
 
@@ -792,18 +792,6 @@ export class ViewportImpl {
     this.fixedLayer_.removeElement(element);
   }
 
-  /** @override */
-  createFixedLayer(constructor) {
-    this.fixedLayer_ = new constructor(
-      this.ampdoc,
-      this.vsync_,
-      this.binding_.getBorderTop(),
-      this.paddingTop_,
-      this.binding_.requiresFixedLayerTransfer()
-    );
-    this.ampdoc.whenReady().then(() => this.fixedLayer_.setup());
-  }
-
   /**
    * Updates touch zoom meta data. Returns `true` if any actual
    * changes have been done.
@@ -1034,10 +1022,7 @@ export class ViewportImpl {
     const oldSize = this.size_;
     this.size_ = null; // Need to recalc.
     const newSize = this.getSize();
-    const promise = this.fixedLayer_
-      ? this.fixedLayer_.update()
-      : Promise.resolve();
-    promise.then(() => {
+    this.fixedLayer_.update().then(() => {
       const widthChanged = !oldSize || oldSize.width != newSize.width;
       this.changed_(/*relayoutAll*/ widthChanged, 0);
       const sizeChanged = widthChanged || oldSize.height != newSize.height;
