+++ src/service/viewport/viewport-interface.js
@@ -316,11 +316,5 @@ export class ViewportInterface extends Disposable {
    * @param {!Element} element
    */
   removeFromFixedLayer(element) {}
-
-  /**
-   * Create fixed layer from constructor (invoked by viewer integration)
-   * @param {typeof ../fixed-layer.FixedLayer} constructor
-   */
-  createFixedLayer(constructor) {}
 }
 /* eslint-enable no-unused-vars */
